app/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
inwx-domrobot==3.0.3
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pyaml==19.4.1
PyYAML==5.1.2
requests==2.22.0
urllib3==1.25.6
uWSGI==2.0.18
Werkzeug==0.16.0
